lab11/django_api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
neo4j==5.17.0
django-cors-headers==4.3.0

